refactor: move all v01 registrations under v01 module
